Remove old peggy based code
