--- wasm-component-model/src/tools/wit.d.ts
@@ -1,8 +1,0 @@
-/* --------------------------------------------------------------------------------------------
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for license information.
- * ------------------------------------------------------------------------------------------ */
-
-import { Document } from './wit-ast';
-
-export function parse(text: string): Document;
